Move issue_registry to homeassistant.helpers
