+++ homeassistant/components/repairs/__init__.py
@@ -2,19 +2,19 @@
 from __future__ import annotations
 
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers.typing import ConfigType
-
-from . import issue_handler, websocket_api
-from .const import DOMAIN
-from .issue_handler import (
-    ConfirmRepairFlow,
+from homeassistant.helpers.issue_registry import (
+    IssueSeverity,
     async_create_issue,
     async_delete_issue,
     create_issue,
     delete_issue,
 )
-from .issue_registry import async_load as async_load_issue_registry
-from .models import IssueSeverity, RepairsFlow
+from homeassistant.helpers.typing import ConfigType
+
+from . import issue_handler, websocket_api
+from .const import DOMAIN
+from .issue_handler import ConfirmRepairFlow
+from .models import RepairsFlow
 
 __all__ = [
     "async_create_issue",
@@ -34,6 +34,5 @@ async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
 
     issue_handler.async_setup(hass)
     websocket_api.async_setup(hass)
-    await async_load_issue_registry(hass)
 
     return True
