+++ homeassistant/components/repairs/models.py
@@ -4,16 +4,7 @@ from __future__ import annotations
 from typing import Protocol
 
 from homeassistant import data_entry_flow
-from homeassistant.backports.enum import StrEnum
 from homeassistant.core import HomeAssistant
-
-
-class IssueSeverity(StrEnum):
-    """Issue severity."""
-
-    CRITICAL = "critical"
-    ERROR = "error"
-    WARNING = "warning"
 
 
 class RepairsFlow(data_entry_flow.FlowHandler):
